fix(pack): improve docs and behaviour
