+++ pack/package-function-app.ts
@@ -1,49 +1,48 @@
+/**
+ * This script packages the Function App in a ZIP archive.
+ *
+ * This way only the neccessary files are included in the archive, where the
+ * func CLI would include all files which are not explicitly ignored.
+ *
+ * In addition it implements a hack neccessary to make the Function App support
+ * ECMAScript modules (ESM): Azure Functions expects ESM entry scripts to have
+ * the extions .mjs, however TypeScript currently can only compile to .js files.
+ *
+ * Therefore, the scriptFiles are renamed to .mjs while packaging.
+ */
+
 import { promises as fs, statSync } from 'fs'
 import path from 'path'
 import os from 'os'
 import { progress, debug } from '../cli/logging.js'
 import { run } from '../cli/process/run.js'
-import chalk from 'chalk'
 import globAsync from 'glob'
 import { promisify } from 'util'
+import { copy, copyFile } from './lib/copy.js'
 const glob = promisify(globAsync)
 
-const copyFile = async (source: string, target: string) => {
-	console.log(
-		chalk.blueBright(source),
-		chalk.magenta('->'),
-		chalk.blueBright(target),
-	)
-	const parts = target.split(path.sep)
-	const targetDir = parts.slice(0, parts.length - 1).join(path.sep)
-	try {
-		await fs.stat(targetDir)
-	} catch {
-		await fs.mkdir(targetDir, { recursive: true })
-	}
-	await fs.copyFile(source, target)
-}
-const copy =
-	(sourceDir: string, targetDir: string) => async (sourceName: string) => {
-		const source = path.resolve(sourceDir, sourceName)
-		const target = path.resolve(targetDir, sourceName)
-		await copyFile(source, target)
-	}
 const packageFunctionApp = async (outZipFileName: string) => {
+	// Don't overwrite an existing file
 	const outFile = path.resolve(process.cwd(), outZipFileName)
 	try {
 		await fs.stat(outFile)
-		throw new Error(`Target file ${outFile} exists.`)
+		console.error(`Target file ${outFile} exists.`)
+		process.exit(1)
 	} catch {
 		// Pass
 	}
+
 	progress('Packaging app', outFile)
 	const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), path.sep))
 	const c = copy(process.cwd(), tempDir)
+
+	// Copy the neccessary files for Azure functions
 	await c('host.json')
+	// ... and for installing dependencies
 	await c('package.json')
 	await c('package-lock.json')
 
+	// Install production dependencies
 	await run({
 		command: 'npm',
 		args: ['ci', '--ignore-scripts', '--only=prod', '--no-audit'],
@@ -52,6 +51,7 @@ const packageFunctionApp = async (outZipFileName: string) => {
 		debug: (info) => debug('[npm]', info),
 	})
 
+	// Find all folder names of functions (they have a function.json in it)
 	const rootEntries = await fs.readdir(process.cwd())
 	const functions = rootEntries
 		.filter((f) => !f.startsWith('.'))
@@ -64,17 +64,15 @@ const packageFunctionApp = async (outZipFileName: string) => {
 			}
 		})
 
+	// Find all compiled JS files, but exclude some development files
+	const excludeDistFolders = ['arm', 'cli', 'pack', 'feature-runner']
 	progress('Packaging app', 'Copying function files')
 	const distJSFiles = await glob(`**/*.js`, {
 		cwd: path.join(process.cwd(), 'dist'),
 	})
-
-	const functionAppFiles = distJSFiles
-		.filter((f) => !f.startsWith('arm/'))
-		.filter((f) => !f.startsWith('cli/'))
-		.filter((f) => !f.startsWith('scripts/'))
-		.filter((f) => !f.startsWith('feature-runner/'))
-
+	const functionAppFiles = distJSFiles.filter(
+		(f) => !excludeDistFolders.includes(f.split(path.sep)[0]),
+	)
 	await Promise.all(
 		functionAppFiles.map(async (f) =>
 			copyFile(
@@ -84,6 +82,8 @@ const packageFunctionApp = async (outZipFileName: string) => {
 		),
 	)
 
+	// Rename the extension of the function handler file to .mjs and update the
+	// function.json
 	await Promise.all(
 		functions.map(async (f) => {
 			const fJSON = path.join(process.cwd(), f, 'function.json')
@@ -105,16 +105,21 @@ const packageFunctionApp = async (outZipFileName: string) => {
 			)
 		}),
 	)
+
+	// ZIP everything
 	await run({
 		command: 'zip',
 		args: ['-r', path.resolve(process.cwd(), outZipFileName), './'],
 		cwd: tempDir,
 		log: (info) => progress('[ZIP]', info),
 	})
+
+	// Remove the temp folder
 	await run({
 		command: 'rm',
 		args: ['-rf', tempDir],
 		log: (info) => progress('Cleanup', info),
 	})
 }
+
 void packageFunctionApp(process.argv[process.argv.length - 1])
